move sample files
